Added functionality to the questions and split it up in components.
The user can now lock in their answer and see if it's correct or incorrect and proceed to the next question.

It's only displaying one question at a time instead of all 10.

Next step is to show the final result when all questions have been answered by navigating to /result.